add documentation and fix visibilities
