+++ src/nl/hannahsten/texifyidea/settings/conventions/TexifyConventionsProjectState.kt
@@ -1,0 +1,18 @@
+package nl.hannahsten.texifyidea.settings.conventions
+
+/**
+ * Per project convention settings.
+ *
+ * The per project settings store a single project scheme and a flag that indicates whether the project scheme or
+ * the global scheme should be used. Instances of this class must be serializable, since they are persisted by a
+ * [TexifyConventionsProjectSettingsManager].
+ */
+internal data class TexifyConventionsProjectState(
+    var scheme: TexifyConventionsScheme = TexifyConventionsScheme(
+        myName = TexifyConventionsScheme.PROJECT_SCHEME_NAME,
+    ),
+    var useProjectScheme: Boolean = false
+) {
+
+    fun deepCopy() = copy(scheme = scheme.deepCopy())
+}
